Moved RecursivePropertyType type and compilation type check
