+++ js/core/index.d.ts
@@ -25,3 +25,30 @@ export type DeepPartial<T> = T extends object ? {
 
 // Omit does not exist in TS < 3.5.1
 export type Skip<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
+
+export type RecursivePropertyType<T, TProps extends string[]> =
+  TProps extends [infer TProp]
+    ? (
+      T extends (infer T2)[]
+        ? RecursivePropertyType<T2, TProps>
+        : (
+          TProp extends string
+            ? (
+              T extends Partial<Record<TProp, infer TValue>>
+                ? TValue
+                : never
+            )
+            : never
+        )
+    )
+    : (
+      TProps extends [infer TProp, ... infer Rest]
+        ? (
+          TProp extends string
+            ? Rest extends string[]
+              ? RecursivePropertyType<RecursivePropertyType<T, [TProp]>, Rest>
+              : never
+            : never
+        )
+        : never
+    );
